# internal/blog/post/infrastructure/handlers/deleteOne.go
package handlers

import (
	"net/http"

	"github.com/CXTACLYSM/hiring-api/internal/blog/post/application/dto"
	"github.com/CXTACLYSM/hiring-api/internal/blog/post/application/services"
	"github.com/CXTACLYSM/hiring-api/pkg/shared/infrastructure/httputils"
	"github.com/CXTACLYSM/hiring-api/pkg/shared/infrastructure/middlewares"
	"github.com/go-chi/chi/v5"
)

type DeleteOnePostHandler struct {
	postService *services.PostService
}

func NewDeleteOnePostHandler(postService *services.PostService) *DeleteOnePostHandler {
	return &DeleteOnePostHandler{
		postService: postService,
	}
}

// ServeHTTP godoc
// @Summary     Delete post
// @Description Soft-deletes a post by ID
// @Tags        posts
// @Security    BearerAuth
// @Param       id   path       string true "Post ID (UUID)"
// @Success     202 "Accepted"
// @Failure     401 {object}    responses.ErrorResponse "Unauthorized"
// @Failure     500 {object}    responses.ErrorResponse "Internal server error"
// @Router      /posts/{id} [delete]
func (h *DeleteOnePostHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	user := middlewares.UserFromRequest(r)
	if user == nil {
		httputils.ResponseError(w, http.StatusUnauthorized, "unauthorized")
		return
	}

	deleteOneDTO := &dto.DeleteOneDTO{
		Id:     chi.URLParam(r, "id"),
		UserId: user.Id,
	}

	if err := h.postService.DeleteOne(deleteOneDTO); err != nil {
		httputils.WriteError(w, err)
		return
	}

	httputils.ResponseOkNoData(w, http.StatusAccepted)
}